Fix "instant" type properties. Add generated unit tests.
Unit tests are generated from the examples present in the FHIR spec. Hence the spec must be downloaded and expanded into downloads/site for the example files to be present so the unit tests can run. The "generate" script will do that but will overwrite the classes (which still have 1-2 hand tweaks in them), so this should improve.

Added a couple of TODOs that point out caveats in the class- and unit test generation.